move visualizer to library (WIP)
Basic library functionality for visualizer achieved! Currently only
builds a static hologram. Lots of code clean-up needing to be taken
care of, and re-implementation of animations is still work to go.
